OnlineStockChecker v2.1
I am calling this version 2 because I already completed a working version last night. The only thing that was added in version 2 was support for email clients to notify the user when a product comes back in stock via email. I am currently trying to set the subject of the email as the product name, then I think this will be 100% complete.

When I am completely done, I will make a separate version of this program without the mail client and with a lot more types of graphic cards across more websites, then I will livestream the live updates.