PublicStockChecker v1.0
This is the first public version of OnlineStockChecker. I'm mostly finished with the personal version of this so I will focus more on the public version. So far this is just a copy of my personal build but with send_email() commented out. I plan on adding an input for the urls, website names, and product names. In the future I hope to automate the website names and product names input based on the urls inputted, but I don't even know if this is possible yet.